Fix: ViewTransform::nonNeg(qint64) for qsizetype counts
Resolves ambiguous nonNeg(qreal) vs nonNeg(int) when subtracting
QList::size() from int totals in gallery size progress.

+++ src/gallerylayout.cpp
@@ -340,7 +340,7 @@ void pack(const QList<ImageItem *> &items, const Params &params,
                 contentW += e.w;
                 contentH = qMax(contentH, e.h);
             }
-            contentW += gap * ViewTransform::nonNeg(row.size() - 1);
+            contentW += gap * ViewTransform::nonNeg(qint64(row.size()) - 1);
             const qreal s = (fill && contentW > 1e-6) ? (layoutW / contentW) : 1.0;
             qreal x = margin;
             qreal placedH = 0.0;
